Summary: Alters flow exponent to 0.68
An examination of the handling of AIM showed that the flow exponent is very important.
There is confusion in the documentation as to whether the value should be 0.68 or 0.611.
It was interpretted that 0.68 was intended, but for some reason 0.611 was used historically.

I examined two blower door test data (my house @ 72 Oceanic; and Shirley Swan's house in New Annan).
72 Oceanic is small/med sized, new, and very tight with vapor barrier.
Shirley Swan's is large sized, old, and has no vapor barrier (i.e. leaky). This house has a factor of 3 higher airflow at 50 Pa.

The data (flow vs Pressure) was plotted and fitted to the equation Q = Co deltaP ^ n.
Interestingly, both house types the n value of 0.68 fit perfectly. The value of 0.61 does not fit well.
So this is definitive.

Thus, the flow coefficient has been updated in the Hse_Gen Script and Template.