Summary: Found two bugs in script and updated 72 Oceanic house
Using the script, I generated the 72 Oceanic house to compare with real life (a good idea).
I found that in the case no windows were specified on a side, it was inadvertantly setting the code to '0' instead of '000000'. This was causing failure and is now corrected.
I also found that the width to depth ratio did not seem to be functioning. I had simply forgot to implement the logic and also messed up the calculation of 'y'. These are both corrected now.

The W/D ratio is correct and is verified for range (0.66 to 1.5).
The Depth is calculated as: D = (Floor_Area / w_d_ratio) ^ 0.5
This was testing and produces the correct aspect ratio.

The foundation was also updated in 72 Oceanic as it is a back walkout, not a full basement.