Summary: AIM-2 Explanation Documents (Helpful!)
As a result of not understanding the ambiguous equations and constants in the existing AIM-2 documents, I did a full evaluation and wrote up the results here.
The document reviews the information from blower door tests, the power law representation (with coefficients and exponents), and how that information is turned into AC/h and ELA (including Bernoulli's equation).

It also provides tables of how to insert the data into a *.aim file of ESP-r as again, this can be confusing.

Lukas


Former-commit-id: e885f1134d8165532bd3fab59b4cca954cb64fe5