Summary: Corrects the template and Hse_Gen script for the orifice discharge coefficient.
There had been confusion as to what the coefficients were. I provided a synopsis document in the previous commit which reviews the blower door data, power law, Cf and n, AC/h and ELA, and orifice discharge coefficient.
From that I determined that because the CSDDRD includes AC/h-50 and ELA-xPa, that I simply needed to specify the orifice discharge coefficient correctly at 0.611.