Summary: Insolation analysis is now conducted for all zones with a tmc file
This commit provides support to conduct a shading/insolation analysis on each zone that has a tmc file.
When Hse_Gen writes the zone info to the cfg file, and encounters a *tmc tag, it will immediately write out 
a:
*isi ./housename.zone.shd

At a later point, obstructions may be added and they will be picked up by ish.

In Sim_Control, the cfg file is now read and any *isi tags have the corresponding zone pushed onto an 
array.
Prior to the bps simulation, ish is run for each zone of the array, creating the insolation files for use 
by bps.