Adds support for occupants.
The number of occupants and the employment ratio is determined from the NN input data.
The ASHRAE fundamentals values are applied to the adults and children and distinct periods (night, morning and evening, day) are setup as gain periods.

The occupants are only applied to the main zones and in proportion to their volumes (bsmt is not included).



Former-commit-id: 139c2ba3278ce1cc8759bc06cfe028b2d8c33fee