Summary: Control strategies for central and distributed thermostat systems.
This commit implements a change in the handling of the control strategies for the houses.
Previously, a basic_5_season controller was specified and each zone referenced it (with the exception of the 
attic/crawl which were free_float).

But to account for centralized system control, a new "slave" type controller was implemented for the CHREM.
In all cases the heating/cooling capability is the volumetric ratio of the heated zones multiplied by the 
heat or cool capacity.
Each zone has an individual controller now, and centralized systems are given slave controller with a main_1 
master.

In testing the system works well - but because the controllers are listed in zone order (bsmt comes first), 
the bsmt slave controller is 1 timestep lagged behind the main_1 master.
This will be corrected in a subsequent commit.



Former-commit-id: a4ff7776b2bd78b7493ee15ae655a741709beddb