Summary: Issue of bsmt slave control lagging the main_1
Code was written to place the main_1 as the first controller in an effort to mitigate the lagged bsmt control 
functionality of the previous commit. Unfortunately, it had no impact. This is because ESP-r solves the zones 
sequentially. So the bsmt (zone 1) was always solved first and would thus always have lagged information from 
zone main_1.

This commit only updates comments and such because the attempt did not work.

I will now examine rezoning such that main_1 is the first zone. This is not trivial as much of my code uses a 
decrement to get at the zone below the operating zone.



Former-commit-id: a02f1c0c9e550324d3390e81939f24828443d727