Summary: Reorganizes zoning so main_1 is zone 1 for master control.
In a previous commit it was recognized the combination of numerically vertical zones (e.g. bsmt is 1) and 
ESP-r consecutive zone solving method led to a slave bsmt controller lagging the main_1 master controller by 
one timestep. 

The only solution to this was to reorder the zones (NOT TRIVIAL).
The numerically vertical zoning was originally chosen to facilitate floor/ceilings facing opposing zones as 
well as air flows to other zones. In the original method, a simple decrement/increment of the zone number was 
sufficient.

However, this is no longer possible because the main_1 must be the master zone for controller purposes of 
houses with central heat (1 thermostat).
So - The code was reworked to zone numerically in the following order: main_X, bsmt, crawl, attic, roof.
Now, when zones are determined originally they are immediately assessed for 'below_zone_name' and 'above...'.

This change actually helped clean up other areas in the code that were not efficient at determining if a zone 
existed below/above.
All areas have been reviewed. The houses have been tested and the results are good.

A comparison of results shows that the bsmt slave controller is now in lock-step with the main_1 master 
controller.