Summary: Solves a thinness problem in houses with no insulation.
In houses with no insulation, I was inserting a 0.1 mm EPS layer to adjust the RSI to fit.
However, this was so thin, and I have a minimum check on the conductivity, that it made not difference.

But when the density and specific heat function was called to adjust for framing, it ended up with a huge 
density because of the thinness.

So now this layer defaults to 5 mm. This seems to have fixed the issue. The layer is thinned by the program 
in an effort to match the RSI. So far it works well.



Former-commit-id: e1710ce6b5464602833b290e05009b6c9d8c7deb