Summary: This commit unfortunately rolls back previous efforts to get the simulation into one file.
Although my previous attempt to mult-thread with a changing of directories was successful, it had the result 
of limiting the simulation computer to using 1 of 8 cores. This obviously is very slow.

This commit rolls back to the two script model which can again take advantage of all simulating cores.
It is similar to r350, but has the benefits of specifying particular house names and running ish on all zones 
for insolation purposes.