Summary: Separates the test/calibration house-type/region from the normals.
It was discovered that when the script was passed '0' for either the house type or the region, it would build all cases including the test or calibration cases.
This was not the intention.

To inhibit this functionality, the test/calibration cases were moved in such a way that a '0' would only catch the intended cases (e.g. SD, DR, AT, QC, OT, PR, BC).