Summary: Adds a kill command for Core_Sim.pl, bps, ish
Because of a previous commit which reinstated the two script type of simulations, there is a lack of control 
over the multi-simulation. In detail, when running Sim_Control.pl, this rapidly spawns the (# of cores) 
instances of 
Core_Sim.pl and then terminates. So you can no longer just kill Sim_Control.pl. Instead you must kill all 
instances of Core_Sim.pl, and then kill instances of bps (usually 8).

This is a simple script that runs pkill (kill by name) for Core_Sim.pl, bps, and ish.