Summary: Adds a Status_Sim.pl file which is used to assess the status of the simulations as they occur.
When the Simulations are occuring it is difficult to assess the condition or status and how things have 
progressed.
This commit changes the control functionality printout such that a different file is created. This is scanned 
by the status checker to determine how much progress and any errors (as determined by the absence of 
dictionary.out).