Summary: Adds tagged format to the reporting to fix some issues. Also comments.
This commit takes the reporting of the status to a tagged format to alleviate some issues with incomplete 
lines.
Previously if a line of data was incomplete it would result in the total being undefined.
This has now been fixed by verifing that sufficent data items exist before overwriting the total information.

Many comments added.



Former-commit-id: 097068d6c5899153b756760812c8217633a60b91