Summary: Fixes two BASESIMP bugs that were noted from the new Status_Sim.pl
1) It appears that the Fortran ESP-r src code is succeptable to number of digits. The fndn height - depth 
must be >= 0.1 m. My code was putting out value of height = 2.2 and depth = 2.10. For some reason (floating 
point issue) this apparently won't work. I had a look at the src code in ESP-r and it does seem coded OK. So 
I have updated my code to be the equivalent of 2.20 and 2.10. We will see if that fixes it.

2) There were conditions where although the house was of reasonable aspect ratio, the bsmt did not cover the 
entire main and this could lead to a huge aspect ratio. So I put a check in on the fndn length/width for the 
minimum/maximum, and would reorganize based on area if required.