Summary: Adds a *end_seconds tag so that we summarize real info
Previously, the status reporting was such that it used the present time, not the end time for calculating 
things such as avg sim.
So if all simulations were complete, this would throw off the numbers badly.

This corrects that and adds reporting of ongoing simulation hours and expected total hours.



Former-commit-id: 12b7df65c324aa2708f55a3ce81cb82b4f3ca457