Summary: Update to the Status report.
The Core_Sim was updated to report the time of each simulation completion.
This is used to get a more accurate avg simulation time, as well as to look at the most recent simulation and see if it is hanging.

The writeout for completed vs still-simulating models has been updated for clarity.