Summary: Alters thin material absorptivity properties to reduce impact.
It was found that exterior facing surfaces with thin materials (such as doors with steel or fiberglass skins) were causing simulation issues.
This was due to their low thermal mass combined with incident solar radiation, which led to a high LW radiation value. ESP-r assumes this value is wrong and uses a correlation, 
resulting in lower LW radiation. Thus the excess flux must becomes stored sensible energy, dramtically raising the temperature of the layer (e.g. 250 C).

At that point the simulator becomes unstable with such large temperature steps and then oscillates back and forth from very cold to very hot (e.g. 5000 C).

The fix is to lower the solar absorptivity, increase the thermal mass, or decrease the timestep.
Decreasing the timestep from 1 hr to 6 minutes works very well. This eliminates the issue of instability, but comes at the expense of modeling time.
Increasing the thermal mass (by increasing the thickness) works to some extent, but deviates from the real multilayer construction.
Decreasing the solar absoptivity works, but again may deviate from reality a bit. Note that increasing the LW emissivity does not help as the correlation will reset the value 
back to some lower value.

In the end, I opted to change the thermal mass of door skins by increasing the thickness from 2 to 3 mm. I also decreased the solar absorptivity from 0.7 to 0.6.
This appears to have solved the problem to some extent, but a check for MZELWE is required for simulations now to make sure that this issue is not still occuring.



Former-commit-id: bef91537a9a5ae66c2dc44ea49419df2a3998591