Summary: Adds support to identify simulation warnings and corrects bad simulation reporting.
The previous commit made changes to help the simulation. Theses were found based on the MZELWE warning.
To aid in subsequent simulation warning detection a scan of the bps file is completed after the simulation to 
identify any warning information. As of right now, the only thing detected is MZELWE, but this may be 
expanded as other warning message indicators are discovered.

The deletion (unlink) of existing summary files related to simulation was added to Sim_Control.pl. 
The read and report of warnings was added to Core_Sim.pl.
Support to report the number of warnings and also a correction to bad_sims was added to Status_Sim.pl.



Former-commit-id: 8cfc1a867917051c14a501e38c53d7842f201bec