Summary: Further increases the thickness of the EPS in non-insulated houses to beat a density problem.
As previously described, the insulation layer is used to account for the thermal mass and specific heat of the framing materials.
In houses without any defined insulation, a minimal insulation layer was added to account for adjusting to the RSI and to account for effects of framing materials.

In a recent commit this insulation layer was increased to 5 mm to offset certain specific heat issues.
During simulation, it was found that although the specific heat issue is solved, the density was still being predicted at SG > 9.
So now this layer has been increased to 10 mm to mitigate this. Again - the conductivity is altered to match the specified RSI. The minimum allowable conductivity has been 
reduced to 5% of that specified in the material database.



Former-commit-id: 595306ce99715c5a59e349d46796d1a43889ccd3