Summary: Using the new reporting, it was discovered that copy&reversed constructions (e.g. attic floor) had a different density and specific heat than the original (e.g. main ceiling).
This was traced to the routine which alters the insulation layer to account for the framing. It was being called a second time to make an adjustment on the reversed
construction.

This has been fixed by checking to see if the construction code is -1, indicating it was reversed.
This logic was also applied to the conductivity adjustment based on RSI, for redundancy purposes.