ummary: fixed 'string' related issue of previous commit, and reduced maximum stone thickness to 0.5 m.
In the previous commit I checked if code == value. I have changed this to if code eq value because the codes can be non-numeric (string).

ESP-r was showing a warning for stone walls over 0.5 m thick. Although HOT2XP specifies the stone wall to be 0.61 m thick, this was altered to the lower value to satisfy ESP-r
Warnings. A quick attempt to change the ESP-r src code did not turn up where the offending value was.