Summary: Adds support for ground source heat pumps.
The Status_Sim found that all simulations with GSHPs (44/16952) were failing. This was traced to a lack of a *.gshp file and a *gshp tag in the *.cfg file.

A review of the required information was conducted and it was found that detailed information is required on the GSHP system. I chose some representative information based on a horizontal system and using example parameters specified in /tester/test_suite/idealized_hvac.

The commit does the following:
1) Adds a template.gshp that is presently a fixed type. This may be updated int the future with more representative information, although there is absolutely no detailed information about the GSHP system in the CSDDRD, simply its presence.

2) The hvac_key.xml was updated because it was found that the backup systems for HP were inappropriate.
For air-source HP with electricity backup - the backup is now an electicity furnace (it previously was distributed baseboards).
For water-source HP with electricity backup - this is now electric boiler (as it is expected that the GSHP is used with a hydronic system).

3) Updates the Control.pm because I was incorrectly formatting for kW when it should have been W. So now 0 places past the decimal instead of 2.

4) Hse_Gen.pl was updated to handle the GSHP template and to decide whether or not it is attributable to the house. In addition cooling is only assumed for ASHP now because GSHP is assumed a hydronic system (no cooling allowed due to condensation). The HP capacities were all updated to 7500 W (fixed) as this is how it is done in HOT2XP. All remaining capacity in the 'heating_capacity' is then attibuted to the Backup heat. This was done in such a way that the minimum amount of backup heat is 1 W. A printout to the hvac file for user readability was corrected so that ASHP, GSHP, and GCEP systems show correctly (it was incorrectly all ASHP before). Finally, the controller capacities were multiplied by 1000 to go from kW to W which is what ESP-r expects in the controller file.