Summary: Performs a range check on longitude difference. ESP-r throws a warning if it is outside -15 to +15 degrees.
There is an instance where the most representative weather station is -15.7 degrees away from the site.
Because it has very minor implications (only a slight results timestamp issue) the value is range checked and the limits are imposed (i.e. it is set back to -15 degrees).
The setback is noted in ../summary_files/Hse_Gen_Issues.txt.