Summary: Fixes very high thermal conductivity values.
The Status_Sim identified construction layers with very high thermal conductivity values (>250).
The occured where the RSI checking algorithm increased the conductivity of the insulation to try and achieve a lower RSI.

The fix:
1) Put a cap on the upper therm conductivity of 250 W/mK (suitable for ESP-r).
2) Found an issue in checking the insulation effects that was causing the further modification to the second layer of insulation, even if the first modification was enough. 
This should now make the RSI values line up very well with the CSDDRD. Also found/fixed logic that did a sum and checked if equal to zero - furthermore this was and RSI calc 
that would set the conductivity to zero, when in fact we wanted maximum (250). Fixed all this.
3) Also added RSI values to each layer reporting (for examination purposes) as well as layer and total construction U-Values (W/m^2K).