Summary: Update of 72 Oceanic ASHP to account for the new values.
By analysis of Kamel Haddad's ASHP (2001) Summary, it appears the rated temperature for ASHP is 8.5 degrees C.
This was determined by plotting the equations for capacity and COP and noting the intersection point with the rated values.

Examining the spec sheet of the 72 Oceanic ASHP, it appears the values should be 10.75 kW and COP = 3.3. There is an additional 16 kW of backup electrical capacity.
So the total heating capacity was set to 26.75 kW.

It is known that the code will intrepet this to be a 7.5 kW ASHP with 19.25 kW backup.