Summary: Provision of the ESP-r format databases and listing of database traits.
In support of Bart L. using CHREM houses under version control (ones that were not generated directly on his 
computer), this commit puts the ESP-r version of the databases under version control.

Usually, these databases are generated with the houses, so as to use the most up to date 
construction/material changes and so there is not redundant information. But that will not work if other 
people are intending to directly use houses (e.g. Bart). So this commit fixes that problem.

Without this commit, running bps on a house in the checked out repository would have errored when looking for 
the existance of the databases.