Summary: HVAC and DHW cleanup
Raises the HP COP from 2 to 3 because ESP-r HVAC file wants to understand the compressor, not the auxiliary 
impacts as well. 2.0 is representative of a complete seasonal combination (and a low one at that).

Corrects an issue that looked in the wrong place for a HP presence when examining the A/C.

Decreases the A/C capacity to be 7.5 kW which is the same as the HP.

Upped the HVAC auxiliarly loads (blowers, pumps) to be more representative.

Added DHW comments

Altered the text of DHW Efficiency to DHW Energy Factor.



Former-commit-id: 49d29a6d56b9a0bc9f47e9392bb2412ea667b29f