Summary: Adds perl.xml which is the highlighting language for text editor kate. This is the latest version (1.24) which supports the '//' operator which is the 'defined-or'. It may replace the version at /usr/share/kde4/apps/katepart/syntax/perl.xml (usually there).
This new operator is being used in the AFN work.
The defined-or operator allows for the assessment of 'definedness' of the left-hand-term, as opposited to 'logical trueness'.

This operator is being used to look up values of hashes, and if the key does not exist, then to die.
This previously would have failed if the value at the key was 0, because that is logical-false.