Summary: Adds Air Flow Network (AFN) support to the CHREM.
This is a significant commit to the CHREM which adds the use of an AFN to support zone-zone and ambient-zone airflows.

There are two reasons to use an AFN:
1) To reduce overheating in the summertime, an ESP-r function (AreWindowsOpen) will verify a number of conditions and if so use the AFN 
instead of AIM-2. Normally, AIM-2 (leakage) overrides AFN, but in cases where the windows are open the AFN prevails. The use of an AFN more 
adequately represents the airflow through these large openings during such periods.

2) The airflow functionality in the *.opr (operations) file is very limited. It only allows 1 infiltration and 1 zone-zone airflow per time 
period. This was no longer suitable for the CHREM, as a main_1 zone must have zone-zone airflow exchange with both the bsmt and main_2 zones. 
This functionality requires an AFN.

To support the use of an AFN, logic was added to Hse_Gen.pl to call functionality in Air_flow.pm. In addition the templates were updated to 
handle the AFN. Finally, certain routines (replace, insert) were moved from Hse_Gen.pl to General.pm so they could be called by other 
modules.

Basically, the AFN places an opening from ambient to the bsmt and main zones whereever there is a window.
Further, a fixed opening is placed between these zones to support stack effect flows.
Finally, a fixed mass flowrate fan is placed between theses zones in two directions (two fans to balance mass) for air circulation purposes.

Additionally, because the the use of *.opr file airflow for zones not listed in the AFN has not been tested by me, I decided to include these 
zones in the AFN as well.

Crawl spaces have fixed openings on all 4 sides with size dependent upon the crawl space type (closed, ventilated, open).
Attic and Roof spaces have fixed openings on all 4 sides (0.25 m^2 per side) as well as eave vents (0.25 m^2 per side, located below the 
attic air point) to support stack effect and mitigate extreme overheating of these uppermost zones during the summertime.

The use of the AFN for the crawl/attic/roof spaces is a significant improvement over fixed infiltration flowrate as imposed by the *.opr 
file.