Summary: Eliminates the opening between zones for natural stack effect.
Although this was a good idea, the interaction between AIM-2 and AFN does not allow for its use.
The AFN will calculate the stack effect assuming the windows are open. The AIM-2 will override the infiltration. But it will not override the zone-zone air flow and thus there will be a mass 
imbalance (although not recognized by the AFN). This will lead to an energy imbalance.

So this functionality was simply disabled and will be left for a future revision that more holistically integrates AIM-2 and AFN.



Former-commit-id: 6a3b994cf8103581d0caa7ae4bf99597275c9ad5