Summary: Simplification and extension of the order subroutine.
Previously, the order subroutine could place some preference. Unfortunately it used a hash to hold indices and as such could not order an array that had two identical values.

This commit does the following:
1) Converts to an array based method where two identical values are allowed.
2) Adds a 'discard' passable array reference to allow for for removing items.

By passing a 'preferred' and 'discard' array references to the routine, it is capable of providing preference and protection to matching items, as well as allowing for the removal of undesireable items.

For example: if an array reference with all zones was passed, the a preferred of (main, bsmt) would place main first, then bsmt. If a discard array reference of ('') was passed, then all remaining items would be discarded.

The extension of this routine is to support new development (such as energy balance in a coming commit) and facilitates things like skipping the BIPV-T zones for energy balance of normal thermal zones.



Former-commit-id: 32b33752207f9c73e39bd5bbd25ecc11c5c0c56a