Summary: updates the input.xml template based on a review of the h3k reports instruction and new CHREM variables.
In reading the h3k reports documentation, it became apparent that many variables are defaulted in the same fashion that I was defining them.
So I converted the values to be non-default and then commented them out. This was you can tell what the alternative is if a default exists.

In a coming ESP-r commit, many CHREM variables were added and this commit takes advantage of that and wildcard specification for finding them.
By using CHREM/* we pick up only those variables defined by us for our use.

The use of log_variable has also been specified for XML output format. This format is easily readable by PERL and can be used to pull in all sorts of summary data and easily modify it.