Summary: A major energy balance summary addition to Core_Sim.pl
It was found that there was great difficulty in getting the ESP-r results to balance. A TON of effort has gone into this and it is now working.
This commit provides the support to analyze the house.xml log file (XML format) and display the integrated fluxes in such a way as to determine energy balance.
The description of the balance will be left to the ESP-r commit (coming shortly), but in general terms shows the following:

An energy balance in created to surround the zone air point and the interior surface nodes.
The conduction, LW radiation (including to surfaces, plant, and casual gains), and SW radiation are summed for each surface by surface type (transparent or opaque).
The infiltration, ambient-zone mechanical ventilation, zone-coupled heat advection, heating system, cooling system, and casual gain convection is summed.
The two preceding items are found to be equal to the stored energy change in the air point, opaque, and transparent node masses.

The results print out in an house.energy_balance file in a vertically organized pattern for easy readibility.