Summary: This commit comments out *.opr airflow and reorders AL gain types
The *.opr airflow is no longer required as this is modeled with an AFN.

The AL gain types were reordered to be more representative of their presence.
Type 20 is AL-Other
Type 21 (Elec) and 22 (NG) are AL-Stove
Type 23 (NG) is AL-Dryer

The AL-Other and AL-Stove manifest as heat within the zone. Although portions may end up outside (e.g. outdoor lighting, stove vent fan), these are very minor at most.
The AL-Dryer ends up entirely outdoors. Although in reality some heat does end up in the zone, this is likely offset by infiltration for dryer.

So only the NG Dryer is added and this is simply for Natural Gas accounting in ESP-r.

Note that DHW is not included as a casual gain. The DHW tank skin/piping losses are accounted for directly by DHW_module.F and MCASI. While the DHW flow in the dwelling likely results in some 
heating, this is offset by bathroom fans and sitting cold water in the dwelling.