Summary: Removes fan power from MVNT.
The Fan power specified in the MVNT file ends up heating the passing air (to some extent) and if it is an HRV (only one implemented), it will pass this information to SiteUtilities 
for summation.

However, we have already accounted for electricity consumption as AL-Other using the NN and specifying a CVS of some type (e.g. HRV).
All of this heat is being injected as a casual gain in the zone.

So this commit zeroes the fan power for the MVNT files so that it is not counted double.
In doing so it will reduce the siteutilities reporting end-uses by one (now only HVAC, DHW, CasualAL).

A similar process should be done for HVAC fans if they have a similar case.



Former-commit-id: e0706bdfb76028214726e5a8f455315bf7bb8f2f