Summary: Removes fan power from HVAC components to avoid doubling up.
As per the previous commit regarding MVNT, it was found that HVAC items were implementing the fan power (say 200 W) into the SiteUtilities via the ventilation functionality. In the
case of furnace and boiler consumption for these items, this was already accounted for in the NN and has become AL-Other and is imposed on the electrical network and as a caual gain.

It was verified that an ASHP was treated as a furnace in the NN, and as such, its circulation fan power was also set to zero. However, the outdoor fan power was maintained (e.g. 250 
W) as this is used by HVAC but does not become ventilation energy consumption.

However, all fan powers (both circulation and outdoor) were left in place for the Air Conditioning. This is because this was not dealt with in the NN. Thus energy required for such a 
circulation fan must be imposed. The HVAC file considers this and then declares the consumption for SiteUtilities under ventilation. This is OK.