Summary: Adds GHG support to the XML reporting in a postprocess way.
Due to the complications of adding data to ESP-r, and that there is no region call in the *.cfg file for 
ESP-r, it was determined the best way to do GHG analysis was to post process. This is successfully working 
now, and is reported in the xml file and summarized in the *.secondary file.

First: GHG is a function of province (not region), so the provincial name was added in a comment in the 
*.cfg file. This makes houses highly succeptable to loss of info if opened/saved in prj, but we knew that 
already.

A GHG_key file was created that stores the GHG multipliers for the range of energy sources (e.g. NG, oil, 
propane), and also includes month by month information for electricity supply and also 
transmission-distribution-efficiency.

A routine was added to XML_reporting which calculates the GHG emissions at each of the detailed levels (say 
end use or energy src type). All energy sources but electricity are treated with an annual emission intensity 
factor and no efficiency factor. Electricity is treated with a monthly factor and includes the 
transmission-distribution efficiency factor (which increases the energy prior to multiplication with the 
intensity factor). The period emissions for electricity are summed from the months as an annual multiplier is 
inappropriate.

Detailed value are stored and after a complete read of the xml file, higher (heirachelly) positions are 
filled with such data. This includes at the site level. So Site GHG is that due to all fuels and then by 
monthly electricity. And the total period value is the sum of the months (again because an annual value is 
inappropriate.

Special consideration is required when making the factors such that the units of the xml reporting and the 
units of the factors are the same. Grams are converted to kg internally.