Summary: Fixes a bug that was inhibiting the integration by month values.
The XML reporting in ESP-r outputs to the log file. For monthly bin normal values (e.g. W) it uses a month 
'index'. For monthly bin integrated values (e.g. GJ) it uses a month 'number'.

This was fixed and the XML files are rewriting correctly with all desireable information. This has made the 
GHG logic work correctly now.

This was verified with a number of test case houses. The electricity and NG GHGs are now calculating 
correctly.