Summary: Alters for better support of AreWindowsOpen.
There is an inherent issue with the use of AreWindowsOpen and timestep selection.
Often, a timestep of 1 hour is used. This will assume that the windows are open for a complete hour and in 
doing so the air change rate will be significant. This will drive the temperature towards ambient. Thus a 
coarse timestep is not appropriate. As the timestep is reduced, this issue goes away.

Basically, temperatures and window opening percentage was set to make things work appropriately. In most 
cases active cooling will be used for houses that have it.
The algorithm for windows opening/closing could be a master's degree in itself.




Former-commit-id: 0b3597030a57d7d62c04e8a433be67161ba917ab