Summary: It was found that due to the seasonal controller that zones were underheated and reaching as low as 10 C for Nova Scotia. This is primarily due the controller entering a free-float spring season mode on April 1, which is to early.
This commit alters these spring/fall periods such that both heating and cooling may be used when the appropriate setpoint is reached.
Although it is recognized that there is a period of time where occupants allow the temperature to range more greatly and turn off their Heat/Cool system, this is very difficult to predict 
and is highly dependent on the climate. Therefore, this has been left out of the present model such that the zone temperatures are more appropriate.



Former-commit-id: 8c5af6fb484a9b98b23a5971dc339f59bf6a6a93