Summary: Fixes a suface-area bug for crawl space type houses.
Unfortunately I coded the SA of aperatures in under a key to the 
aperture name, as opposed to the standard under zone->SA->surface.

This fixes the issue for the crawl space AFN only, and does not fix the 
root of the problem (although it may not ever need to be fixed).



Former-commit-id: 5c7bd63b1a8bab19a7caff9c12076e0fed36f6e1