Summary: Updates the issues reporting to group like items.
Found that the minimum above grade foundation wall height had a variable minimum and was thus printing many issues.
Collated this into one issue that prints out both the value and the variable minimum for each house.

Also corrected an RSI check. We were tripping on 0.01 RSI difference which is obviously actually terrific agreement.
Set so the abs(diff) < 0.1.



Former-commit-id: e28d4cd3b9a0499144531e679439b90259b9fbcf