Summary: Incorporates the use of program 'timelimit' to kill simulations that hang.
There are certain simulations that are hanging (although few) and it is primarily due to MZELWE long wave calculation returning rediculous results because the temperature is off so much 
(e.g. 2000 K or negative K).
Because this hand inhibits subsequent house simulations, a watchdog program is employed, timelimit.

An upper timelimit for bps has been set at 180 seconds (3 minutes) to aid in this prevention. This is obviously only suitable for normal houses of the CHREM at a 1 hour timestep.
When the timestep is reduced or technologies added this value must be modified accordingly.

After a test phase it is likely that I will increase this to 10 minutes to avoid the above conflict (but it is not enough for things like BIPV-T).