Summary: Introduce the V71 folder which contains the CREEM version of the SHEU 1993 houses. The efforts of the CREEDAC team approximately 10 years ago were to populate the SHEU 1993 data with structural information etc. to make a valid HOT2000 file.
I am now planning to have these files read in by a batch processor (through Brian Bradley, NRCan) and the fields output to CSV so that we can use it in the CHREM model.

This commit updates the SHEU 1993 billing data filename. Note that the number of houses in here correspond with Aydinalp's Ch. 3 discussion. There are 8767 houses (although the filenaming goes up to 10979 there are gaps). Of these 2749 contain some or all billing data. Only 1067 contain all billing data. The file says the billing data in GJ or kWh and included Electricity, NG, and Oil. I will be able to figure out which by the export from BB.

Also included in this commit is a file listing the houses without billing data. I plan to use 'rm' to delete these files so that BB only has to run the ones with billing data.


Former-commit-id: 60cc3b07fb9e0f084ce07f52943ff9ca683e1274