Summary: Adds a check for NAN in the data.
Certain houses (approximately 18 of 17000) had nan results in the xml reporting log output. A check of the bps output did not shed any light as to why this was occuring.
To mitigate the issue, a check is now conducted to verify that the data of all desired elements of a house does not have an 'nan' within it.

If it does, the house is stored in the 'bad_houses' category and is deleted from the results. In this fashion it is removed prior to the scaling calculation.



Former-commit-id: bf6b51f4dd3a03f454ee641193f1d4c1cb9478cb